requirements.txt
# Core LLM and AI
anthropic>=0.40.0
openai>=1.54.0
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0

# RAG and Vector Database
chromadb>=0.5.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
sentence-transformers>=3.0.0

# HuggingFace Datasets
datasets>=2.20.0
huggingface-hub>=0.24.0

# PII Anonymisation
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# Web Scraping
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.41.0
aiohttp>=3.10.0
tenacity>=9.0.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.9.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
tiktoken>=0.7.0

# Document handling
pypdf>=4.0.0
markdown>=3.7

# Web Interface
streamlit>=1.38.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Development
black>=24.8.0
isort>=5.13.0
